framework for creating and managing reflists of regression dbs removed; all regression tests now based solely on the query-approach
