--- tests/reflist.json
@@ -1,170 +1,0 @@
-[
-    {
-        "cyclus-ref": "fc5bbee", 
-        "sha1-checksum": "451b951e378c2883c506568ff7763119e23281db", 
-        "input-file": "1_Enrichment_2_Reactor.xml", 
-        "fname": "fc5bbee_8b2d61c_1_Enrichment_2_Reactor.h5", 
-        "cycamore-ref": "8b2d61c"
-    }, 
-    {
-        "cyclus-ref": "fc5bbee", 
-        "sha1-checksum": "e7028d6c70ff6606792d90045d0a3495590e46d9", 
-        "input-file": "2_Sources_3_Reactors.xml", 
-        "fname": "fc5bbee_8b2d61c_2_Sources_3_Reactors.h5", 
-        "cycamore-ref": "8b2d61c"
-    }, 
-    {
-        "cycamore-ref": "v0.4-rc1", 
-        "sha1-checksum": "0060c462e82a5c23db7cf5f10b2b1a1e67462125", 
-        "input-file": "1_Enrichment_2_Reactor.xml", 
-        "fname": "v0.4-rc1_v0.4-rc1_1_Enrichment_2_Reactor.h5", 
-        "cyclus-ref": "v0.4-rc1"
-    }, 
-    {
-        "cycamore-ref": "v0.4-rc1", 
-        "sha1-checksum": "be0d48a5bc229c4d46d8d72de63fa6da4f4097ae", 
-        "input-file": "2_Sources_3_Reactors.xml", 
-        "fname": "v0.4-rc1_v0.4-rc1_2_Sources_3_Reactors.h5", 
-        "cyclus-ref": "v0.4-rc1"
-    }, 
-    {
-        "cyclus-ref": "v0.4", 
-        "sha1-checksum": "3249ab2b874e71aef722e292986a96263b96c6c5", 
-        "input-file": "1_Enrichment_2_Reactor.xml", 
-        "fname": "v0.4_v0.4_1_Enrichment_2_Reactor.h5", 
-        "cycamore-ref": "v0.4"
-    }, 
-    {
-        "cyclus-ref": "v0.4", 
-        "sha1-checksum": "89b930987402f04c84831d8ec4e8b2e61e0116e8", 
-        "input-file": "2_Sources_3_Reactors.xml", 
-        "fname": "v0.4_v0.4_2_Sources_3_Reactors.h5", 
-        "cycamore-ref": "v0.4"
-    }, 
-    {
-        "cycamore-ref": "0.4.1", 
-        "sha1-checksum": "5b747f39b8c8ddd5e26563f8090384eb290b8248", 
-        "input-file": "1_Enrichment_2_Reactor.xml", 
-        "fname": "0.4.1_0.4.1_1_Enrichment_2_Reactor.h5", 
-        "cyclus-ref": "0.4.1"
-    }, 
-    {
-        "cycamore-ref": "0.4.1", 
-        "sha1-checksum": "1aa69d570a672ebff869c04de64e178c7bc4d488", 
-        "input-file": "2_Sources_3_Reactors.xml", 
-        "fname": "0.4.1_0.4.1_2_Sources_3_Reactors.h5", 
-        "cyclus-ref": "0.4.1"
-    }, 
-    {
-        "cyclus-ref": "0.4.2", 
-        "sha1-checksum": "4e95d446a2dc167786845974119584804b3c3eca", 
-        "input-file": "1_Enrichment_2_Reactor.xml", 
-        "fname": "0.4.2_0.4.2_1_Enrichment_2_Reactor.h5", 
-        "cycamore-ref": "0.4.2"
-    }, 
-    {
-        "cyclus-ref": "0.4.2", 
-        "sha1-checksum": "c64b84195f17f5ab68e3efc409044d35b3ce134d", 
-        "input-file": "2_Sources_3_Reactors.xml", 
-        "fname": "0.4.2_0.4.2_2_Sources_3_Reactors.h5", 
-        "cycamore-ref": "0.4.2"
-    }, 
-    {
-        "cycamore-ref": "0.4.3", 
-        "sha1-checksum": "687c40669300671cc7e81cc6f8b3b0e467fc880c", 
-        "input-file": "1_Enrichment_2_Reactor.xml", 
-        "fname": "0.4.3_0.4.3_1_Enrichment_2_Reactor.h5", 
-        "cyclus-ref": "0.4.3"
-    }, 
-    {
-        "cycamore-ref": "0.4.3", 
-        "sha1-checksum": "d7acf9f1ec5bd90059bb38640f3ea6d0d64126f9", 
-        "input-file": "2_Sources_3_Reactors.xml", 
-        "fname": "0.4.3_0.4.3_2_Sources_3_Reactors.h5", 
-        "cyclus-ref": "0.4.3"
-    }, 
-    {
-        "cyclus-ref": "0.4.4", 
-        "sha1-checksum": "e0b9bae64c15d1b3de4204836228b35c06a82fe2", 
-        "input-file": "1_Enrichment_2_Reactor.xml", 
-        "fname": "0.4.4_0.4.4_1_Enrichment_2_Reactor.h5", 
-        "cycamore-ref": "0.4.4"
-    }, 
-    {
-        "cyclus-ref": "0.4.4", 
-        "sha1-checksum": "c179b2297b923664906e77ea23045d4e1c442773", 
-        "input-file": "2_Sources_3_Reactors.xml", 
-        "fname": "0.4.4_0.4.4_2_Sources_3_Reactors.h5", 
-        "cycamore-ref": "0.4.4"
-    }, 
-    {
-        "cycamore-ref": "1.0", 
-        "sha1-checksum": "1ff53f7af4fe9357ee724264b6b895cd6a5c955d", 
-        "input-file": "1_Enrichment_2_Reactor.xml", 
-        "fname": "1.0_1.0_1_Enrichment_2_Reactor.h5", 
-        "cyclus-ref": "1.0"
-    }, 
-    {
-        "cycamore-ref": "1.0", 
-        "sha1-checksum": "41d798d76de4ce6277880c125b058c1bc9a7db94", 
-        "input-file": "2_Sources_3_Reactors.xml", 
-        "fname": "1.0_1.0_2_Sources_3_Reactors.h5", 
-        "cyclus-ref": "1.0"
-    }, 
-    {
-        "cyclus-ref": "1.0.0-rc4", 
-        "sha1-checksum": "2df6acb1aeb8304fea6dbf91f6773dcee4b421e0", 
-        "input-file": "1_Enrichment_2_Reactor.xml", 
-        "fname": "1.0.0-rc4_1.0.0-rc4_1_Enrichment_2_Reactor.h5", 
-        "cycamore-ref": "1.0.0-rc4"
-    }, 
-    {
-        "cyclus-ref": "1.0.0-rc4", 
-        "sha1-checksum": "9de63fb594eeef3da316e73ea3dd05035a476548", 
-        "input-file": "2_Sources_3_Reactors.xml", 
-        "fname": "1.0.0-rc4_1.0.0-rc4_2_Sources_3_Reactors.h5", 
-        "cycamore-ref": "1.0.0-rc4"
-    }, 
-    {
-        "cycamore-ref": "1.1.0", 
-        "sha1-checksum": "ef73b03e4236f707d9b618ffbaa8f3acd715d37e", 
-        "input-file": "1_Enrichment_2_Reactor.xml", 
-        "fname": "1.1.0_1.1.0_1_Enrichment_2_Reactor.h5", 
-        "cyclus-ref": "1.1.0"
-    }, 
-    {
-        "cycamore-ref": "1.1.0", 
-        "sha1-checksum": "5e27c9ad6ba2dd80de0060a986939619e2d5a490", 
-        "input-file": "2_Sources_3_Reactors.xml", 
-        "fname": "1.1.0_1.1.0_2_Sources_3_Reactors.h5", 
-        "cyclus-ref": "1.1.0"
-    }, 
-    {
-        "cyclus-ref": "v1.1.1", 
-        "sha1-checksum": "2218597542aafd19e558580eee696f0faad99c63", 
-        "input-file": "1_Enrichment_2_Reactor.xml", 
-        "fname": "v1.1.1_v1.1.1_1_Enrichment_2_Reactor.h5", 
-        "cycamore-ref": "v1.1.1"
-    }, 
-    {
-        "cyclus-ref": "v1.1.1", 
-        "sha1-checksum": "ec3bf30bad33ec6babc320458aac56910f0ff77a", 
-        "input-file": "2_Sources_3_Reactors.xml", 
-        "fname": "v1.1.1_v1.1.1_2_Sources_3_Reactors.h5", 
-        "cycamore-ref": "v1.1.1"
-    }, 
-    {
-        "cycamore-ref": "v1.2.0", 
-        "sha1-checksum": "d6b317d8bb28197e33af8cdd6f497b7383c84539", 
-        "input-file": "1_Enrichment_2_Reactor.xml", 
-        "fname": "v1.2.0_v1.2.0_1_Enrichment_2_Reactor.h5", 
-        "cyclus-ref": "v1.2.0"
-    }, 
-    {
-        "cycamore-ref": "v1.2.0", 
-        "sha1-checksum": "966d77cb9a62effd5888537a39abb20853c85b62", 
-        "input-file": "2_Sources_3_Reactors.xml", 
-        "fname": "v1.2.0_v1.2.0_2_Sources_3_Reactors.h5", 
-        "cyclus-ref": "v1.2.0"
-    }
-]
