--- tests/test_cases.py
@@ -1,6 +1,0 @@
-
-#List of input files and reference databases
-sim_files = [
-    '../input/physor/1_Enrichment_2_Reactor.xml',
-    '../input/physor/2_Sources_3_Reactors.xml'
-    ]
